requirements.txt
fastapi==0.135.0
uvicorn==0.41.0
sqlalchemy==2.0.47
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
python-dotenv==1.2.1
bcrypt==4.0.1
email-validator==2.3.0
dnspython==2.8.0
pydantic==2.12.5
